requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cssselect2==0.8.0
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.58.2
gunicorn==23.0.0
idna==3.10
mercadopago==2.3.0
packaging==25.0
pillow==11.2.1
psycopg==3.3.2
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.17.2
python-decouple==3.8
pytz==2025.2
requests==2.32.3
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
whitenoise==6.9.0
zopfli==0.2.3.post1
